requirements.txt
fastapi>=0.135.1,<1.0.0
uvicorn
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
httpx
alembic
jinja2
python-multipart
itsdangerous
openai>=1.30.0
pytest>=8.0
pytest-asyncio>=0.24
deepeval>=2.0
